requirements.txt
cycler==0.10.0
drawnow==0.72.0
kiwisolver==1.0.1
matplotlib==3.0.2
numpy==1.16.1
pandas==0.24.1
pkg-resources==0.0.0
pyparsing==2.3.1
pyserial==3.4
python-dateutil==2.8.0
pytz==2018.9
six==1.12.0
